Updated and managed testcases.
